refactor: move function inside utils
